+++ crates/wdl-analysis/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+#### Changed
+
+* Type mismatch diagnostics now distinguish between custom types and references
+  to custom types ([#757](https://github.com/stjude-rust-labs/sprocket/pull/757)).
+* The `UnusedInput` rule no longer ignores `File` and `Array[File]` inputs with specific names ([#741](https://github.com/stjude-rust-labs/sprocket/pull/741)).
+
 #### Fixed
 
 * Added enum type coercion support